Aghiunto autenticazione Aggiunti i controlli tutti servizi(ad esempio prima di  cancellare un utente controlliamo che utente corrente corrisponde a quel utente che deve essere cancellato e cosi via in questa maniera)

+++ src/main/java/com/develhope/spring/services/implementations/AlbumServiceImpl.java
@@ -4,13 +4,16 @@ import com.develhope.spring.dtos.requests.AlbumRequestDTO;
 import com.develhope.spring.dtos.responses.AlbumResponseDTO;
 import com.develhope.spring.entities.Album;
 import com.develhope.spring.entities.Artist;
+import com.develhope.spring.entities.UserEntity;
 import com.develhope.spring.exceptions.EmptyResultException;
 import com.develhope.spring.repositories.AlbumRepository;
 import com.develhope.spring.repositories.ArtistRepository;
 import com.develhope.spring.services.interfaces.AlbumService;
+import com.develhope.spring.utils.Security;
 import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,6 +23,9 @@ import java.util.stream.Collectors;
 
 @Service
 public class AlbumServiceImpl implements AlbumService {
+
+    //TODO CONFIGUR. ENABLE, DISABLE, CREARE RUOLO CHE PUO FARE STE COSE (ADMIN)
+    //todo aggiungere controllo di Current User  anche x i metodi getById
 
     private final ModelMapper modelMapper;
     private final AlbumRepository albumRepository;
@@ -74,9 +80,15 @@ public class AlbumServiceImpl implements AlbumService {
     @Override
     @Transactional
     public AlbumResponseDTO updateAlbum(Long id, AlbumRequestDTO albumRequestDTO) {
+        UserEntity currentUser = Security.getCurrentUser();
+
         Optional<Album> existingAlbumOpt = albumRepository.findById(id);
         if (existingAlbumOpt.isPresent()) {
             Album existingAlbum = existingAlbumOpt.get();
+
+            if (! existingAlbum.getArtist().getUser().equals(currentUser)) //todo exception
+                throw new AccessDeniedException("You are not allowed to update this album");
+
             modelMapper.map(albumRequestDTO, existingAlbum);
             Album updatedAlbum = albumRepository.saveAndFlush(existingAlbum);
             return modelMapper.map(updatedAlbum, AlbumResponseDTO.class);
@@ -88,8 +100,12 @@ public class AlbumServiceImpl implements AlbumService {
     @Override
     @Transactional
     public AlbumResponseDTO deleteAlbumById(Long id) {
+        UserEntity currentUser = Security.getCurrentUser();
 
         return albumRepository.findById(id).map(album -> {
+            if (! album.getArtist().getUser().equals(currentUser))
+                throw new AccessDeniedException("You are not allowed to delete this album");
+
             albumRepository.deleteById(id);
             AlbumResponseDTO deleted = modelMapper.map(album, AlbumResponseDTO.class);
