+++ src/main/java/com/develhope/spring/services/implementations/SongServiceImpl.java
@@ -11,6 +11,7 @@ import com.develhope.spring.repositories.AlbumRepository;
 import com.develhope.spring.repositories.GenreRepository;
 import com.develhope.spring.repositories.PlaylistRepository;
 import com.develhope.spring.repositories.SongRepository;
+import com.develhope.spring.utils.Security;
 import com.develhope.spring.utils.UniversalFieldUpdater;
 import com.develhope.spring.services.interfaces.MinioService;
 import com.develhope.spring.services.interfaces.SongService;
@@ -20,6 +21,7 @@ import org.apache.tomcat.util.http.fileupload.impl.FileSizeLimitExceededExceptio
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,6 +32,9 @@ import java.util.stream.Collectors;
 
 @Service
 public class SongServiceImpl implements SongService {
+
+    //TODO CONFIGUR. ENABLE, DISABLE, CREARE RUOLO CHE PUO FARE STE COSE (ADMIN)
+    //todo aggiungere controllo di Current User  anche x i metodi getById
 
 
     private final SongRepository songRepository;
@@ -85,8 +90,12 @@ public class SongServiceImpl implements SongService {
     @Override
     @Transactional
     public SongResponseDTO updateSong(Long id, SongRequestDTO song) {
+        UserEntity currentUser = Security.getCurrentUser();
 
         return songRepository.findById(id).map(songFound -> {
+            if (! songFound.getAlbum().getArtist().equals(currentUser))// todo ex handler
+                throw new AccessDeniedException("You are not allowed to update this song");
+
             validDtoAndUpdateSong(song, songFound);
 
             Song updated = songRepository.saveAndFlush(songFound);
@@ -171,10 +180,14 @@ public class SongServiceImpl implements SongService {
     @Override
     @Transactional
     public void deleteSongFromMinioStorage(Long songID) {
-       
+        UserEntity currentUser = Security.getCurrentUser();
+
         Song song = songRepository.findById(songID).orElseThrow(
                () -> new EntityNotFoundException("[Song delete failed] Song with ID " + songID + " not Found!")
        );
+
+        if (!song.getAlbum().getArtist().getUser().equals(currentUser))// todo ex handler
+            throw new AccessDeniedException("You are not allowed to delete this song");
         
         String fileToDelete = song.getObjectStorageFileName();
 
@@ -191,6 +204,7 @@ public class SongServiceImpl implements SongService {
     public SongResponseDTO findSongById(long id) {
 
         return songRepository.findById(id).map(songFound -> {
+
             SongResponseDTO response = modelMapper.map(songFound, SongResponseDTO.class);
             songResponseDtoSetArtistName(songFound, response);
 
@@ -202,13 +216,18 @@ public class SongServiceImpl implements SongService {
     @Override
     @Transactional
     public SongResponseDTO deleteSongById(Long id) {
+        UserEntity currentUser = Security.getCurrentUser();
+
         return songRepository.findById(id).map(songFound -> {
+            if (! songFound.getAlbum().getArtist().getUser().equals(currentUser))// todo ex handler
+                throw new AccessDeniedException("You are not allowed to delete this song");
 
             List<Playlist> playlists = playlistRepository.findAll();
             playlists.forEach(playlist -> playlist.getSongs().remove(songFound));
             playlistRepository.saveAll(playlists);
 
             songRepository.deleteById(id);
+            deleteSongFromMinioStorage(id);
 
             SongResponseDTO deleted = modelMapper.map(songFound, SongResponseDTO.class);
             songResponseDtoSetArtistName(songFound, deleted);
@@ -220,7 +239,7 @@ public class SongServiceImpl implements SongService {
 
 
     @Override
-    public void deleteAllSongs() {
+    public void deleteAllSongs() { //todo configurare security per non far calcellare tutto ad artista
         songRepository.deleteAll();
     }
